+++ Packages/com.unity.inputsystem/InputSystem/InputManagerStateMonitors.cs
@@ -14,7 +14,7 @@ namespace UnityEngine.InputSystem
         ////TODO: support combining monitors for bitfields
         public void AddStateChangeMonitor(InputControl control, IInputStateChangeMonitor monitor, long monitorIndex, uint groupIndex)
         {
-            Debug.Assert(m_DevicesCount > 0);
+            if (m_DevicesCount <= 0) return;
 
             var device = control.device;
             var deviceIndex = device.m_DeviceIndex;